Move some consts to const and all data storing classes to a new musiccast_data file
