+++ aiomusiccast/const.py
@@ -101,3 +101,7 @@ ALARM_WEEK_DAYS = [
 
 ALARM_ONEDAY = "oneday"
 ALARM_WEEKLY = "weekly"
+MC_LINK = "mc_link"
+MAIN_SYNC = "main_sync"
+MC_LINK_SOURCES = [MC_LINK, MAIN_SYNC]
+NULL_GROUP = "00000000000000000000000000000000"
